Add event parsers
This commit adds a new event parser, `LoginInitEventParser`, which implements the `IEventParser` interface. The parser is responsible for parsing JSON messages of type `EventType.LoginInit` and extracting the necessary data.

The `Parse` method in the `LoginInitEventParser` class parses the JSON string and retrieves the value of the `DeviceId` property from the `EventData`. If any required properties are missing or have invalid formats, an exception will be thrown.

Additionally, a new file, `LoginInitEventParser.cs`, has been added to the GameServer/Features/Login directory. This file contains the implementation of the `LoginInitEventParser` class.

Note: The code changes in this commit are related to adding support for parsing login initialization events.