Add Constants class for common values
This commit adds a new file `Constants.cs` in the `Common` directory. The file contains two constants: `WebSocketReceiveBufferSize` with a value of 1024 * 4, and `EndpointUrl` with a value of "http://localhost:13371/ws".

In the `Client.cs` file, the WebSocket connection URL is updated to use the `EndpointUrl` constant instead of hardcoding it.

In the `Program.cs` file, the server URL is updated to use the `EndpointUrl` constant.

In the `WebSocketHandler.cs` file, the buffer size for receiving messages is updated to use the `WebSocketReceiveBufferSize` constant.